server/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
aioredis==1.3.1
altgraph==0.17.2
amqp==5.1.1
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
auto-py-to-exe==2.10.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.4.0
blessings==1.6
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
bpython==0.18
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.2.4
canvas==0.3
celery==5.2.6
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
conda==4.3.16
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
cupshelpers==1.0
curtsies==0.3.1
cycler==0.10.0
daphne==3.0.2
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
distro==1.4.0
Django==3.2.8
django-extensions==3.1.5
django-polymorphic==3.1.0
django-ratelimit==3.0.1
django-templated-mail==1.1.1
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==1.16.0
docutils==0.17.1
Eel==0.12.4
elizabeth==0.0.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.2
Fraction==1.5.1
future==0.18.2
GDAL==3.0.4
gevent==21.8.0
gevent-websocket==0.10.1
googletrans==3.0.0
greenlet==1.1.1
grpcio==1.16.1
h11==0.12.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpie==2.5.0
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
ifaddr==0.1.6
IMDbPY==6.8
incremental==21.3.0
ipython==7.13.0
ipython_genutils==0.2.0
isc==2.0
iso8601==1.0.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.15.2
Jinja2==3.0.2
keyring==18.0.1
Kivy==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.2
kombu==5.2.4
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libretranslatepy==2.1.1
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
minecraft==0.1.2
msgpack==1.0.2
multidict==5.1.0
nemo-emblems==5.0.0
netaddr==0.7.19
netifaces==0.10.4
nodeenv==1.6.0
numpy==1.21.2
oauthlib==3.1.1
onboard==1.4.1
outcome==1.1.0
packaging==21.2
PAM==0.4.2
parso==0.5.2
pathtools==0.1.2
pbr==5.8.1
peewee==3.14.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.3.0
ply==3.11
progressbar==2.5
prompt-toolkit==2.0.10
protobuf==3.19.1
proton-client==0.7.1
protonvpn-cli==3.11.1
protonvpn-gui==1.7.0
protonvpn-nm-lib==3.8.0
psutil==5.5.1
psycopg2==2.8.4
psycopg2-binary==2.9.1
py3dns==3.2.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pygame==2.0.1
Pygments==2.10.0
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.2
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
python3-openid==3.2.0
pythondialog==3.4.0
pyttsx3==2.90
pytz==2021.3
pyxdg==0.26
PyYAML==6.0
redis==4.1.2
reportlab==3.5.34
requests==2.26.0
requests-file==1.4.3
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
russian-names==0.1.2
SecretStorage==2.3.1
selenium==3.141.0
serial==0.0.97
Serializer==0.2.1
service-identity==21.1.0
setproctitle==1.1.10
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==1.9.5
sqlparse==0.4.2
ssh-import-id==5.10
stevedore==3.5.0
systemd-python==234
Telethon==1.23.0
tinycss2==1.0.2
tldextract==2.2.1
traitlets==4.3.3
translate==3.6.1
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.9
ufw==0.36
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
vine==5.0.0
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
wadllib==1.3.3
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
wrapt==1.13.3
wsproto==1.0.0
xkit==0.0.0
yarl==1.6.3
youtube-dl==2021.4.26
zope.event==4.5.0
zope.interface==5.4.0
